[refac] 애플리케이션 도커 파일 내용을 infrastructure 계층으로 이동
